refactor: Refactor auth layout
