+++ src/pages/index.js
@@ -1,6 +1,9 @@
 export { default as Browse } from './Browse';
 export { default as CreateEvent } from './CreateEvent';
+export { default as ForgotPassword } from './ForgotPassword';
 export { default as LandingPage } from './LandingPage';
 export { default as Login } from './Login';
+export { default as ResetPassword } from './ResetPassword';
+export { default as ResetPasswordSuccess } from './ResetPasswordSuccess';
 export { default as SignUp } from './SignUp';
 export { default as UserAccount } from './UserAccount';
